Add initial contract test and helpers
Add an initial contract test that covers a new returning teacher that
specifies a teacher reference number, is based in the UK and provides their
telephone number.

As much of the test logic as possible has been extracted into
`ContractHelper` so that the scenarios are as light/easy to read as possible.

The first time the test is ran the contract/output is written. Subsequent
runs test the payload against the current snapshot; if they differ the test
will fail (similar to React snapshot testing).